refactor(server): delete migrated MentorDetails interface, update imports
